Aplicacao_Web/requirements.txt


psycopg2-binary
bcrypt
python-dotenv
openai>=1.0.0
google-genai==2.2.0
pydantic>=2.0
fastapi>=0.110
uvicorn[standard]>=0.27
bert-score
torch
pandas
openpyxl
scipy
scikit-learn
streamlit>=1.35.0
plotly
sshtunnel
paramiko
cryptography
